Move filter code to the processor state module.
